apps/api/requirements.txt
pandas>=1.5.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0
numpy>=1.21.0
datasets>=2.10.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
accelerate>=0.26.0
uvicorn>=0.22.0
